ocrtesseract/ocrai/requirements.txt
numpy<=2.2.6
opencv-python<=4.12.0.88
pandas<=2.3.3
pdf2image<=1.17
pillow<=12.0.0
pytesseract<=0.3.13
openai==0.28.0